Aligns Blue boxes at the top of dashboard. Splits them into their own component to reduce prop drilling.
