+++ src/components/farmInfo/balance/Balance.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState, useContext } from "react";
-import HarvestContext from "../../Context/HarvestContext";
+import HarvestContext from "../../../Context/HarvestContext";
 import styled, { ThemeProvider } from "styled-components";
-import { darkTheme, lightTheme, fonts } from "../../styles/appStyles";
-import harvest from "../../lib/index.js";
+import { darkTheme, lightTheme, fonts } from "../../../styles/appStyles";
+import harvest from "../../../lib/index.js";
 import BalanceSkeleton from "./BalanceSkeleton";
 const { ethers } = harvest;
 
@@ -21,6 +21,7 @@ const BluePanel = styled.div`
   align-items: center;
   justify-content: center;
   flex-direction: column;
+  width: 100%;
   h1 {
     font-size: 2.4rem;
     margin-bottom: 0.5rem;
@@ -30,7 +31,7 @@ const BluePanel = styled.div`
   }
   @media (max-width: 1107px) {
     padding: 2rem 0.7rem 4rem 1.5rem;
-    margin-bottom: 1rem;
+    margin: 1rem 0rem;
 
     h1 {
       font-size: 2.2rem;
@@ -41,8 +42,9 @@ const BluePanel = styled.div`
   }
 `;
 
-const Balance = ({ state }) => {
+const Balance = () => {
   const {
+    state,
     currentExchangeRate,
     prettyBalance,
   } = useContext(HarvestContext);
