+++ src/components/farmInfo/balance/BalanceSkeleton.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled, { ThemeProvider } from "styled-components";
-import { darkTheme, lightTheme } from "../../styles/appStyles";
+import { darkTheme, lightTheme } from "../../../styles/appStyles";
 
 const SkeletonPanel = styled.div`
 display: flex;
@@ -13,6 +13,7 @@ border: ${(props) => props.theme.style.mainBorder};
 box-shadow: ${(props) => props.theme.style.panelBoxShadow};
 border-radius: .5rem;
 overflow: hidden;
+margin-right: 1rem;
 
     .skeleton-header {
         margin: 0;
